plugin: move check_map_for_dne_only ()
Problem: The check_map_for_dne_only () helper function is specific to
flux-accounting in that it iterates through the users map and looks to
see if the plugin is still waiting on flux-accounting data to be
loaded. Thus, it would make more sense for it to be moved to
accounting.*.

Move the definition for check_map_for_dne_only () to accounting.*. Use
auto for the definition of the iterator used in this function.