t1027: add wait-event after updating job
Problem: There is a test in t1027-mf-priority-issue376.t that updates
the duration of a job but does not wait for it to finish updating
before checking its contents in the eventlog, which could cause the
test to fail.

Add a "wait-event" call to wait for the job after it updates before
checking the eventlog.