job_updated (): remove extra Association lookup
Problem: There is a leftover association lookup in the job_updated ()
callback from when the callback was both validating AND performing the
actual update (i.e before a callback was added for
job.update.system.attributes.queue). Thus, this lookup does not need to
be present.

Remove this extra association lookup.