plugin: add project information unpack(s)
Problem: The priority plugin has no way to unpack project information
from the flux-accounting database.

Add two new members to the Association class: projects and def_project,
which will store the projects an association is allowed to charge as
well as their default project.

Add an unpack of the "projects" and "def_project" columns for each
association when unpacking them in the rec_update_cb () function.

Add a new callback to unpack information from the project_table and
store it in a vector of projects to be used for project validation
and annotation.