plugin: rename add_missing_bank_info ()
Problem: The add_missing_bank_info () function is poorly named and not
very reflective of what the function is actually doing.

Rename the function add_special_association ().

Edit the calls to this function to reflect the function name change.

Change the name of the Association variable in the function from "b"
to "a".