accounting: add project validate helper function
Problem: There exists no way to validate a project for an association in
the priority plugin.

Add a new helper function get_project_info () which will validate a
passed-in project for an association. If the project is unknown to
flux-accounting or is an invalid one for an association to submit their
job under, an integer < 0 will be returned, otherwise 0.