plugin: add jobspec-update of project name
Problem: The priority plugin has no way to update jobspec with the
project used for the job in the case where an association submits a job
under a default project.

Add a helper function to the plugin that adds the project name for
an association's job to jobspec via jobspec-update under
attributes.system.project.

Add a jobspec-update of the association's default project name when the
job is in the job.new callback.

In the event where an association submits a job before any information
is loaded to the priority plugin and their job is held in PRIORITY
state, add a jobspec-update of the association's default project name in
job.state.priority if they are submitting under a default project.