plugin: add helper function to add project name
Problem: The priority plugin has no way to update jobspec with the
project used for the job in the case where an association submits a job
under a default project.

Add a helper function to the plugin that adds the project name for
an association's job to jobspec via jobspec-update under
attributes.system.project.