plugin: add jobspec-update of project name
Problem: If a user submits a job under a default project, there is no
way to tell which project they submitted their job under because it does
not get posted to jobspec.

Add a jobspec-update of the association's default project name when the
job is in the job.new callback.

In the event where an association submits a job before any information
is loaded to the priority plugin and their job is held in PRIORITY
state, add a jobspec-update of the association's default project name in
job.state.priority IF their default project name is not "*" ("*" is
the default project name if a user is added to the database with no
projects).