plugin.query: add projects, def_project
Problem: Project information per-association is not returned with the
other flux-accounting information returned by the plugin.query callback.

Add the projects and def_project fields to the list of information
returned for each association in the plugin.query callback.

Add the projects and def_project key-value pairs to the expected files
for t1019-mf-priority-info-fetch.t. Add some projects to one of the
associations in the test and ensure those projects show up when looking
at the ouptut of the plugin.query callback.