accounting: add max_nodes to Association class
Problem: The Association class does not have a max_nodes attribute which
it will need in order for the plugin to enforce a max_nodes limit
per-association.

Add max_nodes as an attribute to the Association class in
accounting.hpp. Add max_nodes values to the initializations of
Association objects for the appropriate unit tests.