plugin: add max_nodes to information payload(s)
Problem: The max_nodes attribute per-association is not sent from the
flux-accounting database to the plugin on a data update.

Add it to the set of information sent over to the plugin. Unpack this
attribute in the callback that unpacks all of the flux-accounting
information per-association. Set a default value for it in the case
where we are creating a special association for someone who submitted a
job *before* the plugin is loaded with flux-accounting information.