repo: add pkg.m4, checks for flux-core libs
Problem: The flux-accounting repo does not check for certain flux-core
libraries that it will eventually need to use, such as the hostlist
library.

Add pkg.m4 to the config/ directory which will enable flux-accounting to
check for flux-core libraries while it is building.

Uncomment the PKG_CHECK_MODULES calls in ax_flux_core.m4 so that
flux-accounting actually checks for these libraries.

Add output to configure.ac to show the location of flux-core's hostlist
library.