plugin: add helper functions for counting nnodes
Problem: The priority plugin needs a way to count the number of nodes an
association is using for their job.

Add two new helper functions to the priority plugin.

extract_nodelist () will be responsible for extracting the "nodelist"
key-value pair out of a job's R.

process_nodelist () is responsible for counting the number of nodes from
this nodelist by using flux-core's hostlist library.

Include flux-core's hostlist library in the plugin and in the plugin's
Makefile.