t1011: add new tests for canceled jobs
Problem: t1011-job-archive-interface does not have any tests that ensure
a user's job-usage value does not get affected by a job that never ran.

Add a basic set of tests in t1011-job-archive-interface.t that submits a
job which gets canceled and ensures that the record does not get added
to flux-accounting's jobs table and thus does not affect a user's
job-usage value.