doc: update job-archive notes
Problem: The Accounting Guide does not detail instructions for the new
job archival method in flux-accounting.

Update the flux-accounting guide with the new instructions on setting up
the new Python script that fetches inactive jobs and stores them in the
flux-accounting DB to be used for job usage and fair share calculation.

Update the code block that shows which commands to set up with
flux-cron.

Update the list of tables present in the flux-accounting database to
account for the new "jobs" table.