job_archive_interface: wrap updates in TRANSACTION
Problem: When flux-accounting needs to do a job usage update for all of
the associations in the association_table, it needs to make a number of
updates to a select number of tables in the DB. If there are a lot of
associations, this can slow down the performance of a number of other,
much simpler commands, such as viewing the information for an
association. This is because the job usage updates makes a separate
transaction for each UPDATE statement performed.

Drop the separate commit() calls from the various helper functions in
job_archive_interface.py and instead wrap the entire set of database
updates in a SQLite transaction, which will only be committed to the DB
after all of the updates are done.