test: move unit test for job usage to sharness
Problem: One of the unit tests in test_job_archive_interface.py attempts
to call update_job_usage(). Since the unit tests also executed within a
transaction, unit test 17 in particular attempts to start a transaction
within a transaction, which isn't supported by SQLite. This test,
however, is simple in nature and can also be tested in sharness.

Move this unit test to t1011-job-archive-interface.t. Make a note in the
test file that this is copied over from the unit test file in
test_job_archive_interface.py.

Re-number the set of tests in test_job_archive_interface.py to account
for the removal of the unit test.