conf.update: add unpack of "age" priority weight
Problem: The priority plugin does not unpack the priority weight
associated with an "age" factor for a job's priority calculation.

Add the "age" factor's associated integer weight to the plugin's
internal map that stores the weights of all of the priority factors by
unpacking it in the callback for conf.update.