plugin: add "age" factor in priority calculation
Problem: The priority plugin does not factor in the age of a job in its
priority calculation.

Add a new callback to the plugin for job.state.sched and store the
timestamp of a job in SCHED state in a map where the key is the job ID
and the value is the timestamp. When the job enters job.priority.get,
the timestamp will be used with the time at that moment to calculate the
"age" of the job, where the older the age, the more of a priority bump
the job gets.

If a job currently accumulating age has its urgency changed to 0, the
age of the job will be saved and no longer grow until the urgency of the
job is changed back to a value > 0.

If a job currently accumulating age has one of its attributes updated,
such as the queue or the bank it is running under, any
previously-accumulated age will be cleared and reset for the job.

When the job enters job.state.inactive, the key-value pair for the job
ID and the timestamp of it entering state SCHED is removed.