t1011: change how to check job usage value
Problem: Checking for an exact job usage value can be challenging when
running tests under valgrind because the value might be ever so
slightly different.

Change the way the job usage value is checked to instead just look that
the job usage value is greater than or equal to the expected value.