plugin: initialize factor weights on plugin load
Problem: The weight of each factor is checked every time the priority of
a job is calculated. It would be cleaner to put this check somewhere
else so it does not need to be repeated.

Create defaults for the weights of each factor used in the priority
calculation of a job. Initialize each weight when the plugin is loaded.
In the callback for conf.update, check to see if these weights have been
configured; if not, continue to use the defaults defined at the top of
the file.