doc: add note about scrub-old-jobs, VACUUM command
Problem: The flux-accounting guide does not make a note of the
scrub-old-jobs command or running VACUUM on the database file if space
needs to be reclaimed.

Add a note to the flux-accounting guide about what the scrub-old-jobs
command does and how to tune it.

Make a note about connecting to the database file and running VACUUM if
we want to clean up and optimize the space occupied by the file.