add_job_records: improve error handling
Problem: The add_job_records() function does not handle the case where
a ResourceSet object cannot be created out of the R fetched for each
job.

Add a try-except block for this, and just skip adding the record in the
case where the creation of the ResourceSet object does not work since
it will not contribute to an association's job usage or fair-share
value.