t: move python unit tests to t/ directory
Problem: The unit tests for the Python bindings for flux-accounting live
in a separate directory from the rest of the test files for the
repository, which can sometimes be tedious to keep track of. They were
also located there before flux-accounting had a "python" directory in
the test directory.

Move the tests to t/python/. Give them slightly shorter names than what
they used to be.