python: rename job_archive_interface.py
Problem: job_archive_interface.py could use a better filename since the
contents of the file have changed. The code in this file primarily
focuses on calculating and updating job usage values for associations in
the flux-accounting database.

Rename job_archive_interface --> job_usage_calculation. Adjust the
various files that list and import this file to account for the name
change.