plugin: initialize plugin w/ cores-per-node count
Problem: The priority plugin does not know about basic system
information it will need in order to enforce a max-cores limit per
association, such as the number of cores on a node.

Add an estimation of a cores-per-node count estimate during the
initialization of the priority plugin by fetching resource.R from the
KVS. Store this estimate in a global variable in the plugin.

Add this estimate to the list of information returned in the
plugin.query callback.