configure: fix dots, add more paths
Problem: The paths for FLUX_IDSET_LIBS/FLUX_IDSET_CFLAGS are not listed
in the output of ./configure. There is also slight misformatting in the
output of the FLUX_CORE_CFLAGS label.

Switch the period and colon at the end of the FLUX_CORE_CFLAGS line.

Add FLUX_IDSET_LIBS and FLUX_IDSET_CFLAGS to the output of ./configure.