t: use "-t" option of view-bank
Problem: There are a couple of instances throughout the testsuite that
call "flux account-shares", but there is a cleaner option to just use
the "-t" optional argument from the "view-bank" command.

Convert the calls from "flux account-shares" to just use "view-bank -t".
Edit the expected files in the testsuite to account for the precision
that "view-bank -t" includes in its output of the database hierarchy.