python: add list-banks command
Problem: There is currently no way to list the current banks defined in
the bank_table in the flux-accounting database, especially in a
convenient format like JSON.

Add a new command to the flux-accounting command suite called
"list-banks". When called, this command will create a JSON object
containing all of the active banks in the bank_table. Passing --inactive
will include *all* banks, even those which have been disabled. The
fields output in the JSON object can be customized with a list of fields
to be included with --fields.