priority-update: send instance owner info
Problem: The instance owner cannot run jobs when the priority plugin is
loaded because they are not a part of the flux-accounting database.
flux-accounting should be able to handle this automatically so that
administrators do not have to manually add instance owner information.

Send instance owner information to the priority plugin in the bulk
update script as another valid user to run jobs under. Give it very high
jobs and nodes limits.