t: update expected output
Problem: Now that instance owner information is sent to the priority
plugin, a few of the tests that look for certain information via the
plugin.query callback also need to be updated since the instance owner
information is now included.

In the tests where users are created with a userid of 1001, change these
to 5001 to avoid potential conflict with the uid of the instance owner.

Change the jq checks to search with select() functions to search via
userid instead of indexing the JSON array by index.

Remove the test_cmp calls with the expected files in
t/expected/plugin_state/ and instead just use jq calls to check for
expected fields.