t: automatically generate test UIDs
Problem: As mentioned in #480, there are a number of tests throughout
the test suite that explicitly define UIDs for users created and used
for various tests in flux-accounting, but these have a potential of
conflicting with system-defined UIDs.

Add definitions for test UIDs in sharness.d/flux-accounting.sh.

Replace the use of explicitly defined UIDs in
t1019-mf-priority-info-fetch.t with the UIDs defined in
flux-accounting.sh.