doc: add note about creating DB before service
Problem: The flux-accounting service depends on the creation of the
flux-accounting DB before starting since it needs to establish a
connection to the DB in order to run "flux account" commands. If this is
done out of order, confusing behavior can result from running the
commands, such as an "attempt to write to a readonly database" error.

Add a note to the flux-accounting guide about making sure the
flux-accounting DB has been created before starting the flux-accounting
service and give an example of the error message that can potentially
occur.