cmd: add export-db as a flux account command
Problem: The "export-db" command is located in a separate file in the
src/cmd/ directory and thus requires a different method of calling it
than the other flux-accounting commands, i.e you need to write "flux
account-export-db" instead of "flux account export-db".

Move this file to the directory containing the other Python bindings for
flux-accounting and add it as a regular command like the other bindings
so that you can just call it like "flux account export-db".

Remove the methods and command line argument that deal with establishing
a connection to the flux-accounting DB since the systemd service will
handle establishing that connection.

Adjust the calls in the sharness test and top-level README for export-db
to account for the change.