t1026: add export-db, pop-db commands to test
Problem: t1026-flux-account-perms.t goes through all of the flux account
commands that require admin privileges to run and makes sure that they
cannot be accessed by regular users, but it does not test the
"export-db" or "pop-db" commands.

Add both of these commands to the test suite to ensure they require
admin privileges to be run.