view-job-records: add "project" filter option
Problem: There is no way to filter jobs by project when looking at job
records in flux-accounting's "jobs" table.

Add a "project" column to the jobs table that will store an optional
project name for each job. Update the schema version of the DB with the
addition of the column name.

Add an option to filter jobs by "project" when adding job records to the
"jobs" table and while looking at job records with the view-job-records
command.