t: update helper script, expected output
Problem: There is a helper script that inserts test job records into the
"jobs" table that needs its INSERT statement to be updated to account
for the addition of the "project" column. There is also an expected
output file that needs to be updated to also account for the new
"Project" column in the output of the view-job-records command.

Update both of these in accordance with the addition of the "project"
column to the "jobs" table in the flux-accounting DB.