JobRecord: remove username from initialization
Problem: The "username" field is automatically populated when userid is
passed into the construction of a JobRecord object, but it is listed as
an attribute during initialization.

Remove "username" from the list of attributes needed to be passed in.

Explicitly state which fields are being assigned to which member of a
JobRecord object instead of labeling them via comments.