queue_table: add max_running_jobs column
Problem: There is no definition for a max running jobs limit for a queue
in the flux-accounting database, but there exists a need to limit the
number of jobs an association can run under a certain queue.

Add a new column to the queue_table in the flux-accounting DB:
max_running_jobs, which limits the number of running jobs an association
can have under a particular queue.

Add max_running_jobs to the set of information sent by the flux
account-priority-update command, unpack it in the priority plugin, and
store it in an attribute of the Queue class.