view-user: make "parsable" spelling consistent
Problem: There are multiple --parsable optional arguments throughout
the flux-accounting command suite, but the spelling for it in the
view-user command is not consistent and is spelled --parseable.
Technically, both spellings are correct, but they should be made
consistent throughout the command suite.

Change the spelling of the --parseable optional argument for view-user
to --parsable.