edit-user: add ability to reset projects list
Problem: Passing -1 for the --projects optional argument in the
edit-user command does not work because the edit_user() function thinks
that -1 is a project name, so it doesn't actually reset the projects
list to just "*".

Add a check if --projects=-1 is passed in and call a new helper
function to just reset the projects list to just "*".

Add a unit test and a sharness test to make sure this works as
expected.