flux-account: add new list-projects cmd
Problem: There currently exists no way to list all of the currently
registered projects in the flux-accounting database.

Add a new command to the flux-accounting command suite that creates a
table of all of the currently registered projects in the
flux-accounting database.

Add a couple sharness tests to t1025-flux-account-projects.t that check
the output of calling list-projects with just the default project
listed and after registering a couple of other projects.