doc: add note about manually loading plugin
Problem: The flux-accounting guide does not detail how to manually load
the priority plugin if the path to the plugin is not configured in the
job-manager config file.

Add a note that the plugin can be loaded manually with "flux jobtap
load". Include the flux account-priority-update command so admins know
to send flux-accounting information to the plugin once it is loaded.