spellcheck: fix spellcheck
Problem: 'make check' does not fail on misspelled words.

Problem: Aspell doesn't like dictionary words that end in numbers.

Problem: The spellcheck script successfully runs an aspell "dry run" but
does so without the dictionary.  Later, when it checks for misspellings,
it is counting lines of output and ignores the return value of aspell.

Add the dictionary to the dry run so it fails there if there are
problems.

Remove IPv4 and IPv6 from the dictionary.

Add "dne", "DNE", and "fairshare" to the dictionary.

Add a typos file to the repository so that certain files are ignored
and certain words are ignored as spelling mistakes.