README: remove configuration section of README
Problem: The top-level README outlines installation instructions for the
background scripts for flux-accounting, but the flux-accounting guide on
ReadTheDocs outlines these steps in more detail.

Remove this section from the top-level README. Place the instructions
for running flux-accounting's unit/sharness tests under a new section of
the README called "Testing".