doc: add database administration page
Problem: The top-level README gives some basic overview of database
administration of the flux-accounting database, but now that
flux-accounting has its own documentation site, it would make more sense
to provide more detailed explanations there.

Move the database administration section to a new page on ReadTheDocs.
Add more details about the structure of the flux-accounting database and
how to interact with it.

Reorganize the top-level index.rst file to better display the contents
of the flux-accounting ReadTheDocs site.

Edit the manual page index to match the orientation of flux-core.