t: edit tests to account for root bank check
Problem: There exists a few unit tests that do not make the assumption
that the flux-accounting database does not enforce a 1-root-bank limit,
but now that there is a check for this in add_bank(), these tests fail.

Edit these tests to account for the check that when a root bank is
added, a root bank does not already exist.