flux-account: get rid of dictionary initialization
Problem: flux-account.py manually creates the dictionary of arguments
that is sent via RPC to the flux-accounting service, but since the
service already unpacks these arguments, flux-account.py can just
create a dictionary of the arguments using vars() and send the entire
dictionary.

Use vars() on the parsed args to create a dictionary before sending
the RPC to the appropriate flux-accounting command.

Move the .split() call of the "fields" optional argument to the
flux-accounting service instead of in flux-account.py.