view-user --parsable: improve output formatting
Problem: Since the formatting of the columns in the output of view-user
--parsable is dependent on the length of the longest column name, the
width of every column is forced to match, which can result in some ugly
screen wrapping even in a relatively small font.

Improve the view-user --parsable formatting option by making the width
of each column dynamic based on the length of each column name instead
of based on the length of the longest column name.