__init__.py.in: define default cols for bank_table
Problem: The default fields for the bank_table are defined in the
list_banks() function, but they will have more than one use, so it makes
sense to define these globally in the accounting Python package.

Create a definition for the names of the columns in the bank_table so
that they can be referenced throughout the Python bindings for
flux-accounting.