__init__.py: define default columns for tables
Problem: The default fields for the relevant tables in the
flux-accounting DB are defined manually throughout the bindings,
so it makes sense to define these globally in the accounting Python
fluxacct.accounting package.

Create a definition for the names of the columns in some of the tables
in __init__.py so that they can be referenced throughout the Python
bindings for flux-accounting.