list-banks: use new common formatter
Problem: The list-banks command defines its own formatting for the
output of the command, but the Python bindings now has its own common
formatter and SQL utilities.

Edit the list-banks command to make use of the new formatter and
external Python utilities. Change the default for the --fields optional
argument to be None if it is not specified.