t: add tests for common formatter
Problem: flux-accounting does not have tests that use the common
formatter in the Python bindings.

Add some unit tests for construction of AccountingFormatter objects and
ensuring the correct amount of data is returned from executing a query.
Test that column names can be fetched from AccountingFormatter objects
and are consistent with the ones defined in the flux-accounting DB.
Test that the correct exceptions are raised when invalid column names
are passed through validation.