t: add tests for common formatter
Problem: flux-accounting does not have tests that use the common
formatter in the Python bindings.

Add some sharness tests for the list-bank command with multiple
customization options for the different fields and getting output as
both a table and as JSON. Add some unit tests for construction of
AccountingFormatter objects and ensuring the correct amount of data
is returned from executing a query.