python: add SQLite utitlity file
Problem: the flux-accounting Python bindings have lots of separate
(and probably duplicate) helper code for validating and interacting
with the SQLite database. It would be cleaner and easier to maintain if
it was centralized in one file that could be referenced throughout the
bindings.

Add a new SQL utility file to the Python bindings that will have helper
functions for interacting with the SQLite database. Add a function to
this file that will validate a custom list of table column names
against a valid list of table column names.