view-jobs: convert job records object to string
Problem: The convert_to_str() function used to convert a list of
JobRecord objects to a string does not actually convert it to a string,
but rather just appends separate strings to a list and then returns a
list. Thus, flux-account.py has to have some special logic for just the
view-job-records command to parse the returned data from the function.

Add a .join() call to the convert_to_str() function to actually
construct a string of all of the job records returned by view_jobs().
Remove the special logic in flux-account.py that iterates through the
list of job records and prints it out since it is no longer needed.