t: view_jobs() --> get_jobs()
Problem: The unit tests for jobs_table_subcommands.py use view_jobs()
for checking the correct amount of job records returned, but get_jobs()
is a better function to call here since it returns a list.

Convert the view_jobs() calls to get_jobs() in the unit tests in
t1006_job_archive.py. Since the return value of get_jobs() does not
include the headers for the job records, reduce all of the expected
lengths of job record lists by 1.