t1006: adjust test description
Problem: One of the descriptions of one of the unit tests in
t1006_job_archive.py is incorrectly stating that one of the function
calls should raise an error, but this is not the case; it should just
return a list of length 0.

Adjust the test description to more accurately explain what the expected
behavior is.