formatter: add BankFormatter subclass
Problem: The AccountingFormatter class has good overarching methods
for printing the results of a query to the flux-accounting database,
but there is some functionality specific to viewing bank information
from the database that could use its own subclass.

Add a new subclass called BankFormatter, which contains unique methods
for viewing bank/user information in hierarchical and parsable formats
with the view-bank command. Remove the helper functions previously
defined in bank_subcommands.py in favor of using this new subclass. Add
a --fields optional argument to allow the user to customize which
columns they want to be returned when looking at a row of information
for a bank.