t: change expected format of "view-bank --users"
Problem: The --users option for the view-bank command lists a lot of
columns for every user row in the association_table, which crowds the
output with potentially unneccessary information when looking at which
users belong to the bank being looked at. The format of this option has
been changed to:

1) be formatted in tables, and
2) only include a couple of columns per user row

Adjust the expected output of the --users optional argument to account
for the use of the BankFormatter subclass.