t: remove single bank info from "view-bank --tree"
Problem: The --tree option for the view-bank command still lists the
default information for the bank being viewed *before* printing the
bank hierarchy, which is not necessary when the user passes the --tree
option. Now that the view_bank() function has been reworked to use the
new BankFormatter subclass, these options are more distinct from one
another, and the --tree option no longer includes this info for the
single bank at the top of the output when the --tree option is passed.

Adjust the expected output in the sharness tests to account for the
removal of the default info for the passed-in bank when calling
view-bank with the --tree option.