t: adjust error message when bank doesn't exist
Problem: There is a test in t1023-flux-account-banks.t that looks for a
specific error message when trying to view a bank that does not exist,
but this error message has changed slightly with the introduction of
the BankFormatter subclass.

Change the expected error message in t1023-flux-account-banks.t to
account for the new message format.