t: add --tree to "view-bank -P" in tests
Problem: The parsable option for viewing bank hierarchies in the
flux-accounting database is meant just to be another option for the
--tree option, but now the -P option requires --tree to also be passed
when calling view-bank. There are multiple tests in the flux-accounting
testsuite that call -P without --tree, and thus, these tests fail.

Add --tree to the view-bank -P calls in the testsuite.