view-user: use formatter class
Problem: The view_user() function manually defines its own print
functionality for viewing information about an association in the
flux-accounting database, but the Python bindings have its own
formatting class that does this.

Create a new subclass of AccountingFormatter called
AssociationFormatter. Adjust the initilization of the
AccountingFormatter to accept a custom error message when the query
fails to find any results.

Adjust the view_user() function to use the new AssociationFormatter
class when formatting the output. Remove the --json optional argument
in favor of printing in JSON by default. Add a --fields optional
argument to allow the user to customize their output, similar to the
list-banks command.