t: change which user is deleted from table
Problem: There is a test in t1023-flux-account-banks.t that makes sure
the default bank of a user is updated when one of their user/bank rows
is deleted from the database by removing it in the test prior, but the
wrong user is deleted.

Change which user is deleted in t1023-flux-account-banks.t. Make the
"grep" check in the subsequent test more robust by checking that all
three of the expected strings show up in the .out file instead of
checking that the .out file contains any whole-word matches for just
one of the patterns searched.