t: edit tests to include custom formats, headers
Problem: The tests for the pop-db and export-db commands are outdated
with the improvements made to export_db_info() - namely, the inclusion
of headers in the .csv file & the customization of output.

Adjust the tests in t1009-pop-db.t and t1016-export-db.t to account for
the changes made to export_db_info().