add-user: make --username and --bank required
Problem: The --username and --bank arguments for add-user should be
required arguments, but they are currently treated as optional
arguments. This means that if someone calls add-user without passing
--username or --bank, the corresponding add_user() function will try
to run the function with None values. Requirements should be added for
both of these arguments at the top level so that the call fails if
either one of these is not present.

Add required=True to both the --username and --bank arguments in
add-user. Specify which arguments are being passed where in
flux-account-service.py. Add two tests to t1007-flux-account-users.t to
make sure errors are raised when either one of these is missing from the
add-user call.